+++ lib/authrocket/credential.rb
@@ -2,10 +2,13 @@ module AuthRocket
   class Credential < Resource
     crud :find, :create, :update, :delete
 
+    belongs_to :auth_provider
     belongs_to :user
 
-    attr :api_key, :credential_type
+    attr :credential_type
+    attr :api_key
     attr :password, :password_confirmation
+    attr :access_token, :provider_user_id, :token_expires_at
 
   end
 end